Make update_kernel check for kernel livepatch package unambiguity
For $KGRAFT being true, the update_kernel testcase is in charge of
installing a kernel livepatch incident's (already released) target kernel.
To this end, it searches the supplied $INCIDENT_REPOs for packages whose
name match the common kernel livepatch package name production and
determines the target kernel version from those. However, in case there are
multiple kernel livepatch packages being provided through the specified
$INCIDENT_REPOs, the result is not really well-defined. Make
prepare_kgraft() report an error in this case. While at it, make it also
report an error if no kernel livepatch package is found at all.

Move the install_lock_kernel() invocation outside the loop over the
$INCIDENT_REPOs to after the aforementionted consistency checks. For
preserving relative order, do the same for the $REMOVE_KGRAFT handling
code.

Note that this patch introduces another subtle, but irrelevant change:
the resp. kernel packages had been searched for and installed with the
kernel livepatch $INCIDENT_REPO being enabled beforehand whereas now, this
will be done when those have been disabled already. However, as kernel
livepatch $INCIDENT_REPOs don't contain any kernel packages, there is no
actual change in functionality in this regard.

+++ tests/kernel/update_kernel.pm
@@ -183,30 +183,43 @@ sub prepare_kgraft {
     install_klp_product;
 
     #add repository with tested patch
+    my $incident_klp_pkg;
+    my @all_pkgs;
     my @repos = split(",", $repo);
     while (my ($i, $val) = each(@repos)) {
-        zypper_call("ar $val kgraft-test-repo-$i");
-
-        my $kversion = zypper_search(q(-s -x kernel-default));
-        my $pkgs     = zypper_search("-s -t package -r kgraft-test-repo-$i");
-
+        my $cur_repo = "kgraft-test-repo-$i";
+        zypper_call("ar $val $cur_repo");
+        my $pkgs = zypper_search("-s -t package -r $cur_repo");
         #disable kgraf-test-repo for while
-        zypper_call("mr -d kgraft-test-repo-$i");
+        zypper_call("mr -d $cur_repo");
 
         fully_patch_system;
         foreach my $pkg (@$pkgs) {
-            my $incident_klp_pkg = is_klp_pkg($pkg);
-            if ($incident_klp_pkg && $$incident_klp_pkg{kflavor} eq 'default') {
-                my $wanted_version = right_kversion($kversion, $incident_klp_pkg);
-                install_lock_kernel($wanted_version);
-                last;
+            my $cur_klp_pkg = is_klp_pkg($pkg);
+            if ($cur_klp_pkg && $$cur_klp_pkg{kflavor} eq 'default') {
+                if ($incident_klp_pkg) {
+                    die "Multiple kernel live patch packages found: \"$$incident_klp_pkg{name}-$$incident_klp_pkg{version}\" and \"$$cur_klp_pkg{name}-$$cur_klp_pkg{version}\"";
+                }
+                else {
+                    $incident_klp_pkg = $cur_klp_pkg;
+                }
             }
         }
 
-        my $pversion = join(' ', map { $$_{name} } @$pkgs);
-        if (check_var('REMOVE_KGRAFT', '1')) {
-            zypper_call("rm " . $pversion);
-        }
+        push @all_pkgs, @$pkgs;
+    }
+
+    if (!$incident_klp_pkg) {
+        die "No kernel livepatch package found";
+    }
+
+    my $kversion       = zypper_search(q(-s -x kernel-default));
+    my $wanted_version = right_kversion($kversion, $incident_klp_pkg);
+    install_lock_kernel($wanted_version);
+
+    if (check_var('REMOVE_KGRAFT', '1') && @all_pkgs) {
+        my $pversion = join(' ', map { $$_{name} } @all_pkgs);
+        zypper_call("rm " . $pversion);
     }
 
     power_action('reboot', textmode => 1);
